Moved emit blocker to ExploitFixes.cpp
