+++ NorthstarDedicatedTest/ExploitFixes.cpp
@@ -5,6 +5,9 @@
 #include "NSMem.h"
 #include "cvar.h"
 
+typedef char(__fastcall* function_containing_emit_t)(uint64_t a1, uint64_t a2);
+function_containing_emit_t function_containing_emit;
+ConVar* sv_cheats;
 ConVar* ns_exploitfixes_log;
 #define SHOULD_LOG (ns_exploitfixes_log->m_Value.m_nValue > 0)
 #define BLOCKED_INFO(s)                                                                                                                    \
@@ -394,7 +397,35 @@ void DoBytePatches()
 	}
 }
 
-void ExploitFixes::LoadCallback(HMODULE unused)
+// https://stackoverflow.com/questions/5661642/c-case-insensitive-first-n-characters-string-comparison
+bool strncmp_insensitive(char const* s1, char const* s2, int n)
+{
+	for (int i = 0; i < n; ++i)
+	{
+		unsigned char c1 = static_cast<unsigned char>(s1[i]);
+		unsigned char c2 = static_cast<unsigned char>(s2[i]);
+		if (tolower(c1) != tolower(c2))
+			return false;
+		if (c1 == '\0' || c2 == '\0')
+			break;
+	}
+	return true;
+}
+
+char function_containing_emit_hook(uint64_t unknown_value, uint64_t command_ptr)
+{
+	char* command_string = *(char**)(command_ptr + 1040); // From decompile
+
+	if (!sv_cheats->m_Value.m_nValue && strncmp_insensitive(command_string, "emit", 5))
+	{
+		spdlog::info("Blocking command \"emit\" because sv_cheats was 0");
+		return 1;
+	}
+	return function_containing_emit(unknown_value, command_ptr);
+}
+
+
+void ExploitFixes::LoadCallback(HMODULE baseAddress)
 {
 	spdlog::info("ExploitFixes::LoadCallback ...");
 
@@ -416,4 +447,9 @@ void ExploitFixes::LoadCallback(HMODULE unused)
 
 	ns_exploitfixes_log =
 		new ConVar("ns_exploitfixes_log", "1", FCVAR_GAMEDLL, "Whether to log whenever ExploitFixes.cpp blocks/corrects something");
+
+	HookEnabler hook;
+	sv_cheats = g_pCVar->FindVar("sv_cheats");
+	ENABLER_CREATEHOOK(
+		hook, (char*)baseAddress + 0x5889A0, &function_containing_emit_hook, reinterpret_cast<LPVOID*>(&function_containing_emit));
 }
